requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
flask==2.0.2
gunicorn==20.1.0
werkzeug==2.0.2
Pillow
